Brings in 232 commits by @KoogieKepler removing STLPort 4.5.3, switching to
v145 (VS 2026) + STDC++20 + /MT static CRT, and rebuilding the proprietary
plugin DLLs (Direct3d9 / Direct3d9_ffp / Direct3d9_vsps / DllExport / DPVS)
from source so the EXE/DLL boundary is STL-ABI aligned.

Empirical validation 2026-05-09: standalone Koogie build reaches char-select
against SWGSource StellaBellum VM at 192.168.1.200:44453. World-load FATAL
on creation/default_pc_equipment.iff is a SWGSource content compat issue
that whitengold v1's source diffs already fix (project_swgsource_iff_mismatch
memory entry); to be ported forward as a separate thematic commit.

Original work by @KoogieKepler — per-commit author metadata preserved via
this merge commit. No squash, no rewrite. Per upstream protocol in
project_swg_source_upstreaming.md, when this base is eventually upstreamed
to SWG-Source/master, attribution to Koogie remains in git blame and log.

Cross-references:
- project_koogie_branches.md (branch contents, prior session validation)
- project_koogie_empirical_validation_2026_05_09.md (today's end-to-end test)
- project_swg_source_upstreaming.md (PR-to-master strategy)

+++ src/engine/client/application/Direct3d9/src/win32/Direct3d9.cpp
@@ -3922,7 +3922,7 @@ inline bool Direct3d9::drawPrimitive()
 		if (ms_alphaBlendEnable)
 			Direct3d9_StateCache::setRenderState(D3DRS_COLORWRITEENABLE, ms_colorWriteEnable);
 		else
-			Direct3d9_StateCache::setRenderState(D3DRS_COLORWRITEENABLE, ms_colorWriteEnable & ~D3DCOLORWRITEENABLE_ALPHA);
+			Direct3d9_StateCache::setRenderState(D3DRS_COLORWRITEENABLE, static_cast<DWORD>(ms_colorWriteEnable) & ~static_cast<DWORD>(D3DCOLORWRITEENABLE_ALPHA));
 		Direct3d9_StateCache::setRenderState(D3DRS_ALPHABLENDENABLE, true);
 		Direct3d9_StateCache::setRenderState(D3DRS_ALPHAREF, static_cast<DWORD>(static_cast<float>(ms_alphaTestReferenceValue) * ms_alphaFadeOpacity.a));
 	}
